+++ lib/services/live_helpers/entry.ex
@@ -147,6 +147,23 @@ defmodule Sorcery.LiveHelpers do
 
 
     # {{{ handle_sorcery({:sorcery, msg}, socket)
+    def handle_sorcery({:sorcery, %{command: :mutation_failed, args: %{error: error, handle_fail: cb}} = msg}, socket) when is_function(cb) do
+      assigns = cb.(error, socket.assigns)
+      socket = Map.put(socket, :assigns, assigns)
+      inner_state = Sorcery.PortalServer.handle_info(msg, socket.assigns.sorcery)
+      {:noreply, assign(socket, :sorcery, inner_state)}
+    end
+    def handle_sorcery({:sorcery, %{command: :mutation_failed, args: %{error: %{reason: reason}}} = msg}, socket) do
+      socket = put_flash(socket, :error, reason)
+      inner_state = Sorcery.PortalServer.handle_info(msg, socket.assigns.sorcery)
+      {:noreply, assign(socket, :sorcery, inner_state)}
+    end
+    def handle_sorcery({:sorcery, %{command: :mutation_success, args: %{data: data, handle_success: cb}} = msg}, socket) when is_function(cb) do
+      assigns = cb.(data, socket.assigns)
+      socket = Map.put(socket, :assigns, assigns)
+      inner_state = Sorcery.PortalServer.handle_info(msg, socket.assigns.sorcery)
+      {:noreply, assign(socket, :sorcery, inner_state)}
+    end
     def handle_sorcery({:sorcery, msg}, socket) do
       inner_state = Sorcery.PortalServer.handle_info(msg, socket.assigns.sorcery)
       {:noreply, assign(socket, :sorcery, inner_state)}