+++ lib/services/portal_server/commands/mutation_failed.ex
@@ -1,0 +1,10 @@
+defmodule Sorcery.PortalServer.Commands.MutationFailed do
+  @moduledoc false
+
+  def entry(_msg, %Sorcery.PortalServer.InnerState{} = inner_state) do
+    inner_state
+  end
+
+
+end
+